[NakamuraOS] Implement API, fix toast, refactor menu
