+++ workspaces/frontend/package.json
@@ -21,6 +21,7 @@
     "@reduxjs/toolkit": "^1.8.5",
     "@tailwindcss/aspect-ratio": "^0.4.2",
     "@tailwindcss/forms": "^0.5.3",
+    "@tanstack/react-query": "^4.19.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^14.4.3",
