+++ workspaces/frontend/src/functions/array/index.ts
@@ -1,0 +1,8 @@
+/**
+ * @since 2022/12/08
+ * @author ThinhHV <thinh@thinhhv.com>
+ * @description description
+ * @copyright (c) 2022 Cloudmana Platform
+ */
+
+export { findObject } from './manipulating-arrays'
