+++ workspaces/frontend/src/menu-items/support.ts
@@ -6,13 +6,15 @@
  */
 
 // assets
-import { ChromeOutlined, QuestionOutlined } from '@ant-design/icons'
-
-// icons
-const icons = {
+import {
   ChromeOutlined,
   QuestionOutlined,
-}
+  AntDesignOutlined,
+  BarcodeOutlined,
+  BgColorsOutlined,
+  FontSizeOutlined,
+  CalculatorOutlined,
+} from '@ant-design/icons'
 
 // ==============================|| MENU ITEMS - SAMPLE PAGE & DOCUMENTATION ||============================== //
 
@@ -22,18 +24,55 @@ const support = {
   type: 'group',
   children: [
     {
+      id: 'util-utilities',
+      title: 'Utilities',
+      type: 'collapse',
+      icon: CalculatorOutlined,
+      children: [
+        {
+          id: 'util-typography',
+          title: 'Typography',
+          type: 'item',
+          url: '/typography',
+          icon: FontSizeOutlined,
+        },
+        {
+          id: 'util-color',
+          title: 'Color',
+          type: 'item',
+          url: '/color',
+          icon: BgColorsOutlined,
+        },
+        {
+          id: 'util-shadow',
+          title: 'Shadow',
+          type: 'item',
+          url: '/shadow',
+          icon: BarcodeOutlined,
+        },
+        {
+          id: 'ant-icons',
+          title: 'Ant Icons',
+          type: 'item',
+          url: '/icons/ant',
+          icon: AntDesignOutlined,
+          breadcrumbs: false,
+        },
+      ],
+    },
+    {
       id: 'sample-page',
       title: 'Sample Page',
       type: 'item',
       url: '/sample-page',
-      icon: icons.ChromeOutlined,
+      icon: ChromeOutlined,
     },
     {
       id: 'documentation',
       title: 'Documentation',
       type: 'item',
       url: 'https://cloudmana.gitbook.io/',
-      icon: icons.QuestionOutlined,
+      icon: QuestionOutlined,
       external: true,
       target: true,
     },
