+++ workspaces/frontend/src/menu-items/utilities.ts
@@ -6,24 +6,7 @@
  */
 
 // assets
-import {
-  AppstoreAddOutlined,
-  AntDesignOutlined,
-  BarcodeOutlined,
-  BgColorsOutlined,
-  FontSizeOutlined,
-  LoadingOutlined,
-} from '@ant-design/icons'
-
-// icons
-const icons = {
-  FontSizeOutlined,
-  BgColorsOutlined,
-  BarcodeOutlined,
-  AntDesignOutlined,
-  LoadingOutlined,
-  AppstoreAddOutlined,
-}
+import { BlockOutlined } from '@ant-design/icons'
 
 // ==============================|| MENU ITEMS - UTILITIES ||============================== //
 
@@ -33,33 +16,11 @@ const utilities = {
   type: 'group',
   children: [
     {
-      id: 'util-typography',
-      title: 'Typography',
+      id: 'util-template',
+      title: 'Templates',
       type: 'item',
       url: '/typography',
-      icon: icons.FontSizeOutlined,
-    },
-    {
-      id: 'util-color',
-      title: 'Color',
-      type: 'item',
-      url: '/color',
-      icon: icons.BgColorsOutlined,
-    },
-    {
-      id: 'util-shadow',
-      title: 'Shadow',
-      type: 'item',
-      url: '/shadow',
-      icon: icons.BarcodeOutlined,
-    },
-    {
-      id: 'ant-icons',
-      title: 'Ant Icons',
-      type: 'item',
-      url: '/icons/ant',
-      icon: icons.AntDesignOutlined,
-      breadcrumbs: false,
+      icon: BlockOutlined,
     },
   ],
 }
