+++ workspaces/frontend/src/models/auth.ts
@@ -1,0 +1,11 @@
+/**
+ * @since 2022/12/08
+ * @author ThinhHV <thinh@thinhhv.com>
+ * @description description
+ * @copyright (c) 2022 Cloudmana Platform
+ */
+
+export interface IToken {
+  accessToken: string
+  refreshToken: string
+}
