+++ workspaces/frontend/src/modules/dashboard/IncomeAreaChart.tsx
@@ -13,7 +13,6 @@ import { useTheme } from '@mui/material/styles'
 
 // third-party
 import ReactApexChart from 'react-apexcharts'
-import { ThemeType } from 'src/models/theme'
 
 // chart options
 const areaChartOptions: any = {
