+++ workspaces/frontend/src/pages/auth/register.tsx
@@ -5,7 +5,9 @@
  * @copyright (c) 2022 Cloudmana Platform
  */
 
-import { ReactElement } from 'react'
+import { ReactElement, useEffect } from 'react'
+import { useSelector } from 'react-redux'
+import { useRouter } from 'next/router'
 import Link from 'next/link'
 
 // material-ui
@@ -16,37 +18,49 @@ import FirebaseRegister from 'src/modules/auth/auth-forms/AuthRegister'
 import AuthWrapper from 'src/modules/auth/AuthWrapper'
 import type { NextPageWithLayout } from 'src/pages/_app'
 import Loadable from 'src/components/Loadable'
+import { RootState } from 'src/state/reducer'
 
 // ================================|| REGISTER ||================================ //
 
-const Register: NextPageWithLayout = Loadable(() => (
-  <AuthWrapper>
-    <Grid container spacing={3}>
-      <Grid item xs={12}>
-        <Stack
-          direction="row"
-          justifyContent="space-between"
-          alignItems="baseline"
-          sx={{ mb: { xs: -0.5, sm: 0.5 } }}
-        >
-          <Typography variant="h3">Sign up</Typography>
-          <Typography
-            component={Link}
-            href="/auth/login"
-            variant="body1"
-            sx={{ textDecoration: 'none' }}
-            color="primary"
+const Register: NextPageWithLayout = Loadable(() => {
+  const router = useRouter()
+  const { token } = useSelector((state: RootState) => state.auth)
+
+  useEffect(() => {
+    if (token) {
+      router.replace('/')
+    }
+  }, [token])
+
+  return (
+    <AuthWrapper>
+      <Grid container spacing={3}>
+        <Grid item xs={12}>
+          <Stack
+            direction="row"
+            justifyContent="space-between"
+            alignItems="baseline"
+            sx={{ mb: { xs: -0.5, sm: 0.5 } }}
           >
-            Already have an account?
-          </Typography>
-        </Stack>
+            <Typography variant="h3">Sign up</Typography>
+            <Typography
+              component={Link}
+              href="/auth/login"
+              variant="body1"
+              sx={{ textDecoration: 'none' }}
+              color="primary"
+            >
+              Already have an account?
+            </Typography>
+          </Stack>
+        </Grid>
+        <Grid item xs={12}>
+          <FirebaseRegister />
+        </Grid>
       </Grid>
-      <Grid item xs={12}>
-        <FirebaseRegister />
-      </Grid>
-    </Grid>
-  </AuthWrapper>
-))
+    </AuthWrapper>
+  )
+})
 
 Register.getLayout = function getLayout(page: ReactElement) {
   return page
