+++ workspaces/frontend/src/state/auth/hooks.ts
@@ -10,14 +10,15 @@ import axios from 'axios'
 import env from '../../config/env'
 import { authSlice } from './index'
 import { AppState } from '../index'
-import { AuthState, AuthStatus, User } from './types'
+import { AuthState, AuthStatus, LoggedInAccount, User } from './types'
 import { useAppDispatch, useAppSelector } from '../hooks'
 import { DependencyList, useEffect, useMemo } from 'react'
 import { setOpenModal } from '../application/actions'
 import { clientApi } from '../../utils/api'
 import { useSetPendingSwitchAcc } from '../application/hooks'
-import { ToastTypes, useToast } from 'src/components/ToastProvider'
-import { useUpdateNotification } from '../notification/hooks'
+import { ToastTypes, useToastMessage } from 'src/hooks/useToastMessage'
+// import { useUpdateNotification } from '../notification/hooks'
+import { ILoginRequest } from 'src/services/auth/auth.type'
 
 export function useAuthState(): AuthState {
   return useAppSelector((state: AppState) => state.auth)
@@ -37,52 +38,52 @@ export function useAuthFetch() {
 
 export function useLogin() {
   const dispatch = useAppDispatch()
+  const [addToast] = useToastMessage()
   const { status: authStatus, token } = useAuthState()
   const setPendingAcc = useSetPendingSwitchAcc()
-  const { addToast } = useToast()
-  const updateNotify: any = useUpdateNotification()
+  // const updateNotify: any = useUpdateNotification()
 
-  const [{ loading, error }, login] = useAsyncFn(async () => {
-    try {
-      setPendingAcc(true)
-      const {
-        data: { token },
-      } = await axios.post<{
-        token: string
-      }>(`${env.SERVER_API}/api/v1/cloudmana/auth/login`, {
-        username: '',
-      })
+  const [{ loading, error }, login] = useAsyncFn(
+    async (data: ILoginRequest) => {
+      try {
+        setPendingAcc(true)
+        const {
+          data: { token, user },
+        } = await axios.post<LoggedInAccount>(`${env.SERVER_API}/api/v1/cloudmana/auth/login`, data)
 
-      setPendingAcc(false)
+        setPendingAcc(false)
 
-      if (token) {
-        dispatch(
-          authSlice.actions.setLoggedIn({
-            token,
-          }),
-        )
-        updateNotify()(window as any).gtag('event', 'login_click', {
-          event_label: 'Login',
-          event_category: 'login_click',
-        })
-        addToast({ type: ToastTypes.SUCCESS, content: 'Connected to account successfully!' })
-      } else {
-        addToast({ type: ToastTypes.ERROR, content: 'Failed to connect to account!' })
+        if (token) {
+          dispatch(
+            authSlice.actions.setLoggedIn({
+              token,
+              user,
+            }),
+          )
+          // updateNotify()(window as any).gtag('event', 'login_click', {
+          //   event_label: 'Login',
+          //   event_category: 'login_click',
+          // })
+          addToast(ToastTypes.SUCCESS, 'Login successfully!')
+          return true
+        } else {
+          addToast(ToastTypes.ERROR, 'Failed to login!')
+        }
+
+        dispatch(setOpenModal(null))
+      } catch (error: any) {
+        if (error?.response && error?.response?.data?.message?.statusCode === 400) {
+          addToast(ToastTypes.ERROR, error.response.data.message.message)
+        } else {
+          addToast(ToastTypes.ERROR, 'Something was wrong!')
+        }
+        dispatch(authSlice.actions.setNotLoggedIn(AuthStatus.DEACTIVATE))
+        dispatch(setOpenModal(null))
       }
-
-      dispatch(setOpenModal(null))
-    } catch (error: any) {
-      if (error?.response && error?.response?.data?.message?.statusCode === 400) {
-        addToast({ type: ToastTypes.ERROR, content: error.response.data.message.message })
-      } else if (error.code === 4001) {
-        addToast({ type: ToastTypes.ERROR, content: 'User denied message signature!' })
-      } else {
-        addToast({ type: ToastTypes.ERROR, content: 'Transaction has been failed!' })
-      }
-      dispatch(authSlice.actions.setNotLoggedIn(AuthStatus.DEACTIVATE))
-      dispatch(setOpenModal(null))
-    }
-  }, [dispatch, authStatus, token])
+      return false
+    },
+    [dispatch, authStatus, token],
+  )
 
   return {
     login,
